+++ rest-assert-core/src/main/java/com/github/mjeanroy/rest_assert/internal/data/Cookie.java
@@ -76,5 +76,5 @@ public interface Cookie {
 	 *
 	 * @return Max age (in seconds).
 	 */
-	int getMaxAge();
+	long getMaxAge();
 }
