+++ rest-assert-core/src/main/java/com/github/mjeanroy/rest_assert/internal/data/bindings/asynchttp/AsyncHttpCookie.java
@@ -84,7 +84,7 @@ public class AsyncHttpCookie implements Cookie {
 	}
 
 	@Override
-	public int getMaxAge() {
+	public long getMaxAge() {
 		return cookie.getMaxAge();
 	}
 }
