+++ rest-assert-core/src/main/java/com/github/mjeanroy/rest_assert/internal/data/bindings/javax/JavaxCookie.java
@@ -84,7 +84,7 @@ public class JavaxCookie implements Cookie {
 	}
 
 	@Override
-	public int getMaxAge() {
+	public long getMaxAge() {
 		return cookie.getMaxAge();
 	}
 }
