+++ rest-assert-core/src/test/java/com/github/mjeanroy/rest_assert/tests/TestData.java
@@ -49,7 +49,7 @@ public final class TestData {
 		return httpResponse;
 	}
 
-	public static Cookie newCookie(String name, String value, String domain, String path, int maxAge, boolean secured, boolean httpOnly) {
+	public static Cookie newCookie(String name, String value, String domain, String path, long maxAge, boolean secured, boolean httpOnly) {
 		Cookie cookie = mock(Cookie.class);
 		when(cookie.getName()).thenReturn(name);
 		when(cookie.getValue()).thenReturn(value);
