+++ rest-assert-unit/src/test/java/com/github/mjeanroy/rest_assert/api/cookie/CookieAssert_assertHasMaxAge_Test.java
@@ -63,7 +63,7 @@ public class CookieAssert_assertHasMaxAge_Test extends AbstractCookieTest {
 		};
 	}
 
-	protected Cookie cookie(int maxAge) {
+	protected Cookie cookie(long maxAge) {
 		return newCookie("name", "value", "domain", "path", maxAge, true, true);
 	}
 }
